refactor saga
