+++ app/containers/HomePage/saga.js
@@ -1,7 +1,6 @@
 // Gets Yoda-Speak translation of text from MaShape
 
-import { take, call, put, select, takeLatest } from 'redux-saga/effects'
-import { LOCATION_CHANGE } from 'react-router-redux'
+import { call, put, select, takeLatest } from 'redux-saga/effects'
 import { LOAD_TRANSLATION } from 'containers/App/constants'
 import { translationLoaded, translationLoadingError } from 'containers/App/actions'
 
@@ -24,18 +23,9 @@ export function * getTranslation () {
 }
 
 // Root saga manages watcher lifecycle
-export function * translationData () {
+export default function * translationData () {
   // Watches for LOAD_TRANSLATION actions and calls getTranslation when one comes in.
   // By using `takeLatest` only the result of the latest API call is applied.
   // It returns task descriptor so we can continue execution
   yield takeLatest(LOAD_TRANSLATION, getTranslation)
-
-  // Suspend execution until location changes
-  yield take(LOCATION_CHANGE)
-  // yield cancel(watcher)
 }
-
-// Bootstrap sagas
-export default [
-  translationData
-]
